+++ markwon-html/src/main/java/io/noties/markwon/html/tag/SimpleTagHandler.java
@@ -28,6 +28,11 @@ public abstract class SimpleTagHandler extends TagHandler {
 
     @Override
     public void handle(@NonNull MarkwonVisitor visitor, @NonNull MarkwonHtmlRenderer renderer, @NonNull HtmlTag tag) {
+        // @since $nap; check if tag is block one and visit children
+        if (tag.isBlock()) {
+            visitChildren(visitor, renderer, tag.getAsBlock());
+        }
+
         final Object spans = getSpans(visitor.configuration(), visitor.renderProps(), tag);
         if (spans != null) {
             SpannableBuilder.setSpans(visitor.builder(), spans, tag.start(), tag.end());